make reusable TeamsDropdown
