+++ src/components/TeamsDropdown/index.jsx
@@ -8,25 +8,13 @@ import {
   FormHelperText,
   Select,
 } from '@material-ui/core';
-import { setTeams } from '../../../../store/slices/teamsSlice';
-import { getTeamsAllRequestData } from '../../../../services/teamsService';
-import makeFetch from '../../../../services';
-import useStylesMain from '../../../../hooks/useStylesMain';
-import useStylesLocal, { getStyleMenuItem } from './style';
+import { setTeams } from '../../store/slices/teamsSlice';
+import { getTeamsAllRequestData } from '../../services/teamsService';
+import makeFetch from '../../services';
+import useStylesMain from '../../hooks/useStylesMain';
+import useStylesLocal, { getStyleMenuItem, MenuProps } from './style';
 
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      width: 250,
-    },
-  },
-};
-
-const SelectTeam = ({ team_id, value, onChange, error, helperText }) => {
+const TeamsDropdown = ({ team_id, value, onChange, error, helperText }) => {
   const classesMain = useStylesMain();
   const classesLocal = useStylesLocal();
   const theme = useTheme();
@@ -84,4 +72,4 @@ const SelectTeam = ({ team_id, value, onChange, error, helperText }) => {
   );
 };
 
-export default SelectTeam;
+export default TeamsDropdown;
