+++ src/components/TeamsDropdown/style.js
@@ -13,5 +13,17 @@ const getStyleMenuItem = (name, personName, theme) => ({
       : theme.typography.fontWeightBold,
 });
 
-export { getStyleMenuItem };
+const ITEM_HEIGHT = 48;
+const ITEM_PADDING_TOP = 8;
+
+const MenuProps = {
+  PaperProps: {
+    style: {
+      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
+      width: 250,
+    },
+  },
+};
+
+export { getStyleMenuItem, MenuProps };
 export default useStylesLocal;
